% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar-chart.R
\name{bar_chart}
\alias{bar_chart}
\title{Function for plotting beautiful barcharts in ggplot}
\usage{
bar_chart(data, y, x, na.rm = FALSE, title = NULL, sub.title = NULL,
  flip = FALSE, y.title = NULL, x.title = NULL, decreasing = NULL,
  bar.colour.name = NULL, y.lim = NULL, y.breaks = NULL,
  bar.width = NULL, ...)
}
\arguments{
\item{data}{data frame containing data for plotting}

\item{y}{character string specifying column name in data that should be y variable}

\item{x}{character string specifying column name in data that should be x variable}

\item{na.rm}{a logical indicating whether NA values should be stripped
before the computation proceeds}

\item{title}{character string specifying chart title}

\item{sub.title}{character string specifying chart sub title}

\item{flip}{logical indicating whether barchart should be flipped or not}

\item{y.title}{character string specifying y-axis title}

\item{x.title}{character string specifying x-axis title}

\item{decreasing}{logical indicating whether y values should be ordered and in which direction}

\item{bar.colour.name}{character string indicating which bars to be coloured based on x names (see details)}

\item{y.lim}{numeric vector with length two providing y-axis limits (min and max)}

\item{y.breaks}{numeric vector specifying where ticks appear}
}
\description{
Function for plotting beautiful barcharts in ggplot
}
\details{
The option bar.colour.name allows for multiple bars based on x-axis names to get a different colour.
If set to NULL all bars will have same default colour. Since ggplot2 version 2.2.0 this has changed
as the ordering gets complicated when y values contain both positive and negative values. As a result
bar colouring is only allowed decreasing is set to TRUE.
}
\examples{
data(mtcars)
mtcars$name <- rownames(mtcars)
bar_chart(mtcars, "mpg", "name", flip = TRUE,
title = "Miles per gallon across different car models",
scale.y = c(0, 40, 5),
decreasing = TRUE)
}

