% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box-chart.R
\name{box_chart}
\alias{box_chart}
\title{Create box chart}
\usage{
box_chart(df, x, group = NULL, flip = FALSE, title = NULL,
  y.title = NULL, x.title = NULL, ...)
}
\arguments{
\item{df}{data frame containing data for plotting}

\item{x}{character string specifying column name in df that should be x variable}

\item{group}{character string setting the group variable in the data.frame}

\item{flip}{logical indicating whether box chart should be flipped or not}

\item{title}{character string specifying chart title}

\item{y.title}{character string specifying y-axis title}

\item{x.title}{character string specifying x-axis title}
}
\description{
This function creates an overlay plotting a bar chart in ggplot
}

