% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candle-chart.R
\name{candle_chart}
\alias{candle_chart}
\title{Function for plotting beautiful candle-charts in ggplot}
\usage{
candle_chart(data, na.rm = FALSE, title = NULL, sub.title = NULL,
  y.title = NULL, x.title = NULL, x.interval = NULL, lwd = 0.2,
  bar.width = 1, dc.lines = "both", dc.window = 10,
  output.length = NULL, ...)
}
\arguments{
\item{data}{data frame containing data for plotting}

\item{na.rm}{a logical indicating whether NA values should be stripped
before plotting}

\item{title}{character string specifying chart title}

\item{sub.title}{character string specifying chart sub title}

\item{y.title}{character string specifying y-axis title}

\item{x.title}{character string specifying x-axis title}

\item{x.interval}{numeric used to set number of breaks when scaling the x variable}

\item{dc.lines}{character specifying if "both", "none", the "high", or the "low" Donchian
channel line should be used}

\item{dc.window}{integer specifying the with of the time window used in the Donchian
channel calculation}

\item{output.length}{integer. Specifying the number of output observations}
}
\description{
Function for plotting beautiful candle-charts in ggplot
}
\details{
No further comments...
}
\examples{
n <- 30
data <- data.frame(open = rnorm(n), high = rnorm(n)+1, low = rnorm(n)-1, close = rnorm(n))
rownames(data) <- Sys.Date() + 1:n
candle_chart(data, na.rm = T, title = "Example of candle-chart - might look funny due to random numbers",
             sub.title = NULL, y.title = "prices", dc.lines = "both", base.size = 14, x.interval = 5)
}

