% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grey-theme.R
\name{grey_theme}
\alias{grey_theme}
\title{Function for designing a beautiful grey theme for ggplot2}
\usage{
grey_theme(legend.position = "bottom", plot.margin = c(0.7, 1.2, 0.5, 0.5))
}
\arguments{
\item{legend.position}{character string specifying the legend position
(see details)}
}
\description{
Function for designing a beautiful grey theme for ggplot2
}
\details{
The ggplot2 package comes with four options for setting the
legend position - "left", "right", "bottom" and "top"
}

