% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line-chart.R
\name{line_chart}
\alias{line_chart}
\title{Create line chart}
\usage{
line_chart(df, y, x, group = NULL, title = NULL, sub.title = NULL,
  x.title = NULL, y.title = NULL, vline = NULL, hline = NULL,
  x.interval = NULL, date.format = NULL, min.lim = NULL, max.lim = NULL,
  lower.ribbon = NULL, upper.ribbon = NULL, legend.names = NULL,
  ribbon.names = NULL, ...)
}
\arguments{
\item{df}{data frame containing data for plotting}

\item{y}{specifying column name in df that should be y variable}

\item{x}{specifying column name in df that should be x variable}

\item{group}{specifies the label for splitting data into multiple lines}

\item{title}{character string specifying chart title}

\item{sub.title}{character string specifying chart sub title}

\item{x.title}{character string specifying x-axis title}

\item{y.title}{character string specifying y-axis title}

\item{vline}{numeric specifying position of vertical line}

\item{hline}{numeric specifying position of horizontal line}

\item{x.interval}{numeric used to set breaks when scaling the x variable}

\item{date.format}{character vector specifying the date label (see details)}

\item{min.lim}{numeric setting the minimum value on y-axis}

\item{max.lim}{numeric setting the maximum value on y-axis}

\item{lower.ribbon}{character string specifying column with values for lower ribbon}

\item{upper.ribbon}{character string specifying column with values for upper ribbon}

\item{legend.names}{character vector setting custom legend names}

\item{ribbon.names}{character string specifying custom legend name for ribbon}
}
\description{
This function is a convinient overlay for creating a beautiful line
plot using ggplot2
}
\details{
Format codes for time series classes are defined in strftime. If date.format is
set to NULL the date format will be set to the most used format for each time
series class.

The ellipsis is used to pass on arguments to the grey_theme function. Primary
use is to specify the legend.position to either "left", "right", "bottom", "top" or "none".
}

