% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save-png.R
\name{save_png}
\alias{save_png}
\title{Save ggplots into multiple png files}
\usage{
save_png(...)
}
\description{
This function takes ggplots and save them into png images with fixed size
}
\details{
The function saves png files with filename equal to the name of the ggplot object.
The png image is saved with a height of 480 pixels and width scaled by the golden ratio to 777
(776.66 but rounded to nearest integer)
}
\examples{
g1 <- density_chart(mtcars, "mpg", "cyl")
mtcars$name <- rownames(mtcars)
g2 <- bar_chart(mtcars, "mpg", "name", flip = T)
save_png(g1,g2)
}

