% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter-chart.R
\name{scatter_chart}
\alias{scatter_chart}
\title{Create scatter chart}
\usage{
scatter_chart(df, x, y, x.names = NULL, x.names.show = NULL, title = NULL,
  sub.title = NULL, y.title = NULL, x.title = NULL, x.min = NULL,
  x.max = NULL, y.min = NULL, y.max = NULL, ...)
}
\arguments{
\item{df}{data frame containing data for plotting}

\item{x}{character string specifying name of x variable in data frame}

\item{y}{character}

\item{x.names}{character string specifying column name of names for
textual annotations}

\item{x.names.show}{numeric indicating how many point names should be shown (details)}

\item{title}{character string specifying chart title}

\item{sub.title}{character string specifying chart sub title}

\item{y.title}{character string specifying y-axis title}

\item{x.title}{character string specifying x-axis title}

\item{x.min}{numeric setting minimum limit on x-axis}

\item{x.max}{numeric setting maximum limit on x-axis}

\item{y.min}{numeric setting minimum limit on y-axis}

\item{y.max}{numeric setting maximum limit on y-axis}

\item{transparency}{numeric setting the level of colour transparency}
}
\description{
This function is a convinient overlay for creating a beautiful scatter
plot using ggplot2
}
\details{
The x.names.show is based on distance between actual and fitted values.
If x.names.show is set to 3 then the three observations with the largest
absolute distance between actual and fitted values will have their name
shown next to the point.
}
\examples{
density_chart(mtcars, "mpg", "cyl", title = "Miles per gallon",
sub.title = "(per groups of cylinders)", vline = TRUE)
}

