% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array2df.R
\name{array2df}
\alias{array2df}
\title{Converts a 1D array into a data.frame with a "names" column}
\usage{
array2df(array.in)
}
\arguments{
\item{array.in}{one-dimensional array}
}
\description{
Converts a 1D array into a data.frame with a "names" column
}
\details{
Returns a dataframe with a "names" column and a "data" column
}

