% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar-chart.R
\name{bar_chart}
\alias{bar_chart}
\title{Function for plotting beautiful barcharts in ggplot}
\usage{
bar_chart(df, y, x, title = NULL, sub.title = NULL, flip = FALSE,
  y.title = NULL, x.title = NULL, decreasing = NULL,
  bar.colour.name = NULL, scale.y = NULL, ...)
}
\arguments{
\item{df}{data frame containing data for plotting}

\item{y}{character string specifying column name in df that should be y variable}

\item{x}{character string specifying column name in df that should be x variable}

\item{title}{character string specifying chart title}

\item{sub.title}{character string specifying chart sub title}

\item{flip}{logical indicating whether barchart should be flipped or not}

\item{y.title}{character string specifying y-axis title}

\item{x.title}{character string specifying x-axis title}

\item{decreasing}{logical indicating whether y values should be ordered and in which direction}

\item{bar.colour.name}{character string indicating which bars to be coloured based on x names (see details)}

\item{scale.y}{numeric vector with length three providing y-axis limits (min and max) and breaks (see details)}
}
\description{
Function for plotting beautiful barcharts in ggplot
}
\details{
The option bar.colour.name allows for multiple bars based on x-axis names to get a different colour.
If set to NULL all bars will have same default colour.

The scale.y numeric vector has length three and indicates the y-axis limits and break points.
An input vector of c(0, 60, 10) would translate into y-axis with minimum value at 0 and
maximum value of 60 with breaks for every 10 between the minimum and maximum values.

Otherwise see example for clarification on the scale.y variable.
}
\examples{
mtcars$name <- rownames(mtcars)
bar_chart(mtcars, "mpg", "name", flip = TRUE,
title = "Miles per gallon across different car models",
scale.y = c(0, 40, 5))
}

