% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar-chart.R
\name{bar_chart}
\alias{bar_chart}
\title{Function for plotting beautiful barcharts in ggplot}
\usage{
bar_chart(df, y, x, flip = FALSE, title = NULL, y.title = NULL,
  x.title = NULL, ...)
}
\arguments{
\item{df}{data frame containing data for plotting}

\item{y}{character string specifying column name in df that should be y variable}

\item{x}{character string specifying column name in df that should be x variable}

\item{flip}{logical indicating whether barchart should be flipped or not}

\item{title}{character string specifying chart title}

\item{y.title}{character string specifying y-axis title}

\item{x.title}{character string specifying x-axis title}
}
\description{
Function for plotting beautiful barcharts in ggplot
}
\examples{
mtcars$name <- rownames(mtcars)
bar_chart(mtcars, "mpg", "name", flip = TRUE,
title = "Miles per gallon across different car models")
}

