% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density-chart.R
\name{density_chart}
\alias{density_chart}
\title{Create density chart}
\usage{
density_chart(data, x, group = NULL, title = NULL, sub.title = NULL,
  legend.names = NULL, y.title = NULL, x.title = NULL,
  transparency = 0.3, min.lim = NULL, max.lim = NULL, vline = FALSE,
  vline.custom = NULL, ...)
}
\arguments{
\item{data}{data frame containing data for plotting}

\item{x}{character string specifying name of x variable in data frame}

\item{group}{character string for grouping of x}

\item{title}{character string specifying chart title}

\item{sub.title}{character string specifying chart sub title}

\item{legend.names}{character string specifying legend names}

\item{y.title}{character string specifying y-axis title}

\item{x.title}{character string specifying x-axis title}

\item{transparency}{numeric setting the level of colour transparency}

\item{min.lim}{numeric setting minimum limit on x-axis}

\item{max.lim}{numeric setting maximum limit on x-axis}

\item{vline}{logical value for drawing vertical lines for the distribution median of x}

\item{vline.custom}{numeric setting position on vertical line not tied to the distribution of x}
}
\description{
This function is a convinient overlay for creating a beautiful density
plot using ggplot2
}
\details{
Group variable not ideally be a factor but the function will force class factor on the variable.

Use legend names variable to set custom labels on legends. If legend names is NULL the function
will force legend names to the colnames of either x or names of groups depending on the input.

The ellipsis is used to pass on arguments to the grey_theme function. Primary
use is to specify the legend.position to either "left", "right", "bottom" and "top".

Vline option specifies whether vertical lines showing the median of the distribution
should be plotted.
}
\examples{
density_chart(mtcars, "mpg", "cyl", title = "Miles per gallon",
sub.title = "(per groups of cylinders)", vline = TRUE)
}

