% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density-chart.R
\name{density_chart}
\alias{density_chart}
\title{Function for plotting beautiful barcharts in ggplot}
\usage{
density_chart(df, x, group = NULL, title = NULL, y.title = NULL,
  x.title = NULL, transparency = 0.3, min.lim = NULL, max.lim = NULL,
  vline = FALSE, ...)
}
\arguments{
\item{df}{data frame containing data for plotting}

\item{x}{specifying column name in df that should be x variable}

\item{title}{character string specifying chart title}

\item{y.title}{character string specifying y-axis title}

\item{x.title}{character string specifying x-axis title}

\item{y}{specifying column name in df that should be y variable}

\item{flip}{logical indicating whether barchart should be flipped or not}
}
\description{
Function for plotting beautiful barcharts in ggplot
}

