% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density-chart.R
\name{density_chart}
\alias{density_chart}
\title{Create density chart}
\usage{
density_chart(df, x, group = NULL, title = NULL, y.title = NULL,
  x.title = NULL, transparency = 0.3, min.lim = NULL, max.lim = NULL,
  vline = FALSE, vline.custom = NULL, ...)
}
\arguments{
\item{df}{data frame containing data for plotting}

\item{x}{specifying column name in df that should be x variable}

\item{title}{character string specifying chart title}

\item{y.title}{character string specifying y-axis title}

\item{x.title}{character string specifying x-axis title}

\item{y}{specifying column name in df that should be y variable}

\item{flip}{logical indicating whether barchart should be flipped or not}
}
\description{
This function is a convinient overlay for creating a beautiful density
plot using ggplot2
}
\details{
The ellipsis is used to pass on arguments to the grey_theme function. Primary
use is to specify the legend.position to either "left", "right", "bottom" and "top".

Vline option specifies whether vertical lines showing the median of the distribution
should be plotted.
}
\examples{
density_chart(mtcars,"mpg","cyl",title="Miles per gallon",vline=TRUE)
}

