% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line-chart.R
\name{line_chart}
\alias{line_chart}
\title{Create line chart}
\usage{
line_chart(df, y, x = NULL, group = NULL, ...)
}
\arguments{
\item{df}{data frame containing data for plotting}

\item{y}{specifying column name in df that should be y variable}

\item{x}{specifying column name in df that should be x variable}

\item{flip}{logical indicating whether barchart should be flipped or not}

\item{title}{character string specifying chart title}

\item{y.title}{character string specifying y-axis title}

\item{x.title}{character string specifying x-axis title}
}
\description{
This function is a convinient overlay for creating a beautiful line
plot using ggplot2
}
\details{
The ellipsis is used to pass on arguments to the grey_theme function. Primary
use is to specify the legend.position to either "left", "right", "bottom" and "top".
}

