% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line-chart.R
\name{line_chart}
\alias{line_chart}
\title{Create line chart}
\usage{
line_chart(data = NULL, y, x = NULL, group = NULL, title = NULL,
  sub.title = NULL, x.title = NULL, y.title = NULL, vline = NULL,
  hline = NULL, x.interval = NULL, date.format = NULL, y.min = NULL,
  y.max = NULL, ribbon.lwr = NULL, ribbon.upr = NULL,
  legend.names = NULL, ribbon.names = NULL, aspect.ratio = 1.61, ...)
}
\arguments{
\item{data}{data frame containing data for plotting}

\item{y}{specifying column name in data that should be y variable}

\item{x}{specifying column name in data that should be x variable}

\item{group}{specifies the label for splitting data into multiple lines}

\item{title}{character string specifying chart title}

\item{sub.title}{character string specifying chart sub title}

\item{x.title}{character string specifying x-axis title}

\item{y.title}{character string specifying y-axis title}

\item{vline}{numeric specifying position of vertical line}

\item{hline}{numeric specifying position of horizontal line}

\item{x.interval}{numeric used to set number of breaks when scaling the x variable}

\item{date.format}{character vector specifying the date label (see details)}

\item{y.min}{numeric setting the minimum value on y-axis}

\item{y.max}{numeric setting the maximum value on y-axis}

\item{ribbon.lwr}{character string specifying column with values for lower ribbon}

\item{ribbon.upr}{character string specifying column with values for upper ribbon}

\item{legend.names}{character vector setting custom legend names}

\item{ribbon.names}{character string specifying custom legend name for ribbon}
}
\description{
This function is a convinient overlay for creating a beautiful line
plot using ggplot2
}
\details{
Format codes for time series classes are defined in strftime. If date.format is
set to NULL the date format will be set to the most used format for each time
series class.

The ellipsis is used to pass on arguments to the plot_theme function. Primary
use is to specify the legend.position to either "left", "right", "bottom", "top" or "none".
Arguments can also be sent to plot_theme function to adjust plot margins - see function
for more details.
}

