% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter-chart.R
\name{scatter_chart}
\alias{scatter_chart}
\title{Create scatter chart}
\usage{
scatter_chart(data, x, y, na.rm = FALSE, name.label = NULL,
  name.labels.show = NULL, label.just = NULL, title = NULL,
  sub.title = NULL, y.title = NULL, x.title = NULL, x.min = NULL,
  x.max = NULL, y.min = NULL, y.max = NULL, prediction.interval = FALSE,
  ...)
}
\arguments{
\item{data}{data frame containing data for plotting}

\item{x}{character string specifying name of x variable in data frame}

\item{y}{character}

\item{na.rm}{a logical indicating whether NA values should be removed}

\item{name.label}{character string specifying column name of names for
textual annotations}

\item{name.labels.show}{numeric or character string indicating either the
number of label names to show or the specific names (details)}

\item{label.just}{numeric vector with length two setting the horizontal
and the vertical text alignment for data point labels}

\item{title}{character string specifying chart title}

\item{sub.title}{character string specifying chart sub title}

\item{y.title}{character string specifying y-axis title}

\item{x.title}{character string specifying x-axis title}

\item{x.min}{numeric setting minimum limit on x-axis}

\item{x.max}{numeric setting maximum limit on x-axis}

\item{y.min}{numeric setting minimum limit on y-axis}

\item{y.max}{numeric setting maximum limit on y-axis}

\item{prediction.interval}{logical indicating whether a prediction
interval should be computed and ploted}
}
\description{
This function is a convinient overlay for creating a beautiful scatter
plot using ggplot2
}
\details{
If the name.labels.show is numeric the data point labels shown are based
on the distance between actual and fitted y-values. If the variable is set
to 3 then the three points with largest absolute distance between actual
y-values and the fitted values will have their name shown. This functionality
requires that the prediction.interval is set to TRUE.
If the name.labels.show is a character string then names are explicitly shown
in the scatter plot.
}
\examples{
data(mtcars)
density_chart(mtcars, "mpg", "cyl", title = "Miles per gallon",
sub.title = "(per groups of cylinders)", vline = TRUE)
}

